clean up pd-extended auto-build script and prepped the release script
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@9717 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ scripts/auto-build/pd-extended-release-auto-builder.sh
@@ -10,26 +10,26 @@ SCRIPT=`echo $0| sed 's|.*/\(.*\)|\1|g'`
 
 BUILD_DIR=.
 case $SYSTEM in 
-	 linux)
-		  BUILD_DIR=linux_make
-		  echo "Configuring to use $BUILD_DIR on GNU/Linux"
-		  ;;
-	 darwin)
-		  BUILD_DIR=darwin_app
-		  echo "Configuring to use $BUILD_DIR on Darwin/Mac OS X"
-		  ;;
-	 mingw*)
-		  BUILD_DIR=win32_inno
-		  echo "Configuring to use $BUILD_DIR on MinGW/Windows"
-		  ;;
-	 cygwin*)
-		  BUILD_DIR=win32_inno
-		  echo "Configuring to use $BUILD_DIR on Cygwin/Windows"
-		  ;;
-	 *)
-		  echo "ERROR: Platform $SYSTEM not supported!"
-		  exit
-		  ;;
+	linux)
+		BUILD_DIR=linux_make
+		echo "Configuring to use $BUILD_DIR on GNU/Linux"
+		;;
+	darwin)
+		BUILD_DIR=darwin_app
+		echo "Configuring to use $BUILD_DIR on Darwin/Mac OS X"
+		;;
+	mingw*)
+		BUILD_DIR=win32_inno
+		echo "Configuring to use $BUILD_DIR on MinGW/Windows"
+		;;
+	cygwin*)
+		BUILD_DIR=win32_inno
+		echo "Configuring to use $BUILD_DIR on Cygwin/Windows"
+		;;
+	*)
+		echo "ERROR: Platform $SYSTEM not supported!"
+		exit
+		;;
 esac
 
 
@@ -39,12 +39,20 @@ auto_build_root_dir=`pwd`
 echo "root: $auto_build_root_dir" 
 
 # let rsync handle the cleanup with --delete
-rsync -a --delete rsync://128.238.56.50/distros/pd-extended-release/ \
-	 ${auto_build_root_dir}/
+case $SYSTEM in
+	mingw*)
+		/c/cygwin/bin/sh -c \
+			"rsync -a --delete rsync://128.238.56.50/distros/pd-extended/ ${auto_build_root_dir}/"
+		;;
+	*)
+		rsync -a --delete rsync://128.238.56.50/distros/pd-extended/ ${auto_build_root_dir}/
+		;;
+esac
 
 cd "${auto_build_root_dir}/packages/$BUILD_DIR"
-make -C "${auto_build_root_dir}/packages" patch
+make -C "${auto_build_root_dir}/packages" set_version
 make test_locations
+mount
 make package_clean
 make install && make package
 
@@ -54,39 +62,49 @@ upload_build ()
     platform_folder=$1
     build_folder=$2
     archive_format=$3
-
-	 archive="${auto_build_root_dir}/packages/${platform_folder}/${build_folder}/Pd*.${archive_format}"
+	
+	archive=`ls -1 ${auto_build_root_dir}/packages/${platform_folder}/${build_folder}/Pd*.${archive_format} | tail -1`
     
     echo "upload specs $1 $2 $3"
     echo "Uploading $archive"
-	 upload_filename=`ls -1 ${archive} | sed "s|.*/\(.*\)\.${archive_format}|\1-${HOSTNAME}.${archive_format}|"`
-	 case $SYSTEM in 
-		  mingw*)
-				test -e ${archive} && /c/cygwin/bin/sh -c \
-					 "rsync -a ${archive} rsync://128.238.56.50/upload/${DATE}/${upload_filename}" &&\
-					 echo SUCCESS
+	upload_filename=`ls -1 ${archive} | sed "s|.*/\(.*\)\.${archive_format}|\1-${HOSTNAME}.${archive_format}|"`
+	if [ -e ${archive} ]; then
+		case $SYSTEM in 
+			mingw*)
+				/c/cygwin/bin/sh -c \
+					"rsync -a ${archive} rsync://128.238.56.50/upload/${DATE}/${upload_filename}" && \
+					md5sum ${archive} > ${archive}.md5 && \
+					/c/cygwin/bin/sh -c \
+					"rsync -a ${archive}.md5 rsync://128.238.56.50/upload/${DATE}/${upload_filename}.md5" && \
+					echo SUCCESS
 				;;
-		  *)
-				test -e ${archive} && rsync -a ${archive} \
-					 rsync://128.238.56.50/upload/${DATE}/${upload_filename}  && \
-					 echo SUCCESS
+			*)
+				rsync -a ${archive} rsync://128.238.56.50/upload/${DATE}/${upload_filename}  && \
+					md5sum ${archive} > ${archive}.md5 && \
+					rsync -a ${archive}.md5 rsync://128.238.56.50/upload/${DATE}/${upload_filename}.md5  && \
+					echo SUCCESS
 				;;
-		  esac
+		esac
+	fi
 }
 
 
 case $SYSTEM in 
-	 linux)
-		  upload_build linux_make build tar.bz2
-		  ;;
-	 darwin)
-		  upload_build darwin_app . dmg
-		  ;;
-	 mingw*)
-		  upload_build win32_inno Output exe
-		  ;;
-	 cygwin*)
-		  upload_build win32_inno Output exe
-		  ;;
+	linux)
+		if [ -x /usr/bin/dpkg-deb ]; then
+			upload_build linux_make . deb
+		else
+			upload_build linux_make build tar.bz2
+		fi
+		;;
+	darwin)
+		upload_build darwin_app . dmg
+		;;
+	mingw*)
+		upload_build win32_inno Output exe
+		;;
+	cygwin*)
+		upload_build win32_inno Output exe
+		;;
 esac
